requirements.txt
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
filelock==3.0.12
GDAL==3.1.3
gdown==3.12.2
geojson==2.5.0
geomet==0.2.1.post1
html2text==2020.1.16
idna==2.10
iniconfig==1.0.1
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.2
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rpy2==3.3.6
scikit-learn==0.23.2
scipy==1.5.2
sentinelPot==0.1.0
sentinelsat==0.14
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.50.0
tzlocal==2.1
urllib3==1.25.10
